requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.2
django-autoslug==1.9.9
Pillow==9.5.0
pycodestyle==2.10.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.3
